--- src/ls/mod.rs
@@ -1,13 +1,0 @@
-//! Helpers and type definitions to list MP3 files and MP3s in directories.
-//!
-//! This module contains basic methods to list and compare (for sorting) MP3 files from the local filesystem.
-
-mod cmp;
-mod error;
-mod info;
-mod list;
-
-pub use cmp::*;
-pub use error::*;
-pub use info::*;
-pub use list::*;
